1. 기존의 SSO_URL말고 default_url이라는 XE 메인 사이트 정보를 입력받게 해서 가상사이트와 기본 사이트와의 연계를 보다 명확히 구분함. --- default_url은 관리자 > 우상단 config에서 설정하며 이값이 없으면 가상사이트와 기본 사이트를 구분해야 할때 동작을 중지하고 경고를 노출하도록 변경 2. 위젯들의 가상사이트 연관된 부분들을 모두 재검토하여 문제 없도록 수정 --- 글보관함, 달력, 포럼형, 탭형은 모듈이 선택되어 있지 않으면 에러 문구를 출력하도록 변경 3. 가상사이트와 본사이트와의 링크를 XE 앞단에서 처리하여 redirect하는 부분을 정교히 수정

+++ faceOff/classes/widget/WidgetHandler.class.php
@@ -68,7 +68,7 @@
 
             // 캐시 파일을 갱신하여야 할 경우 lock파일을 만들고 캐시 생성
             $oWidget = WidgetHandler::getObject($widget);
-            if(!$oWidget) return;
+            if(!$oWidget || !method_exists($oWidget,'proc')) return;
 
             $widget_content = $oWidget->proc($args);
             FileHandler::writeFile($cache_file, $widget_content);
